+++ revproxy/views.py
@@ -8,7 +8,12 @@ import logging
 
 import urllib3
 
-from django.utils.six.moves.urllib.parse import urlparse, urlencode, quote_plus
+try:
+    from django.utils.six.moves.urllib.parse import (
+        urlparse, urlencode, quote_plus)
+except ImportError:
+    # Django 3 has no six
+    from urllib.parse import urlparse, urlencode, quote_plus
 
 from django.shortcuts import redirect
 from django.views.generic import View
@@ -23,7 +28,7 @@ from .utils import normalize_request_headers, encode_items
 # Chars that don't need to be quoted. We use same than nginx:
 #   https://github.com/nginx/nginx/blob/nginx-1.9/src/core/ngx_string.c
 #   (Lines 1433-1449)
-QUOTE_SAFE = '<.;>\(}*+|~=-$/_:^@)[{]&\'!,"`'
+QUOTE_SAFE = r'<.;>\(}*+|~=-$/_:^@)[{]&\'!,"`'
 
 
 ERRORS_MESSAGES = {
@@ -43,6 +48,7 @@ class ProxyView(View):
     _upstream = None
     cert_required = True
 
+    add_x_forwarded = False
     add_remote_user = False
     default_content_type = 'application/octet-stream'
     retries = None
@@ -121,22 +127,45 @@ class ProxyView(View):
         """Return request headers that will be sent to upstream.
 
         The header REMOTE_USER is set to the current user
-        if the view's add_remote_user property is True
+        if AuthenticationMiddleware is enabled and
+        the view's add_remote_user property is True.
 
         .. versionadded:: 0.9.8
+
+        If the view's add_x_forwarded property is True, the
+        headers X-Forwarded-For and X-Forwarded-Proto are set to the
+        IP address of the requestor and the request's protocol (http or https),
+        respectively.
+
+        .. versionadded:: TODO
 
         """
         request_headers = self.get_proxy_request_headers(self.request)
 
-        if self.add_remote_user and self.request.user.is_active:
+        if (self.add_remote_user and hasattr(self.request, 'user')
+                and self.request.user.is_active):
             request_headers['REMOTE_USER'] = self.request.user.get_username()
             self.log.info("REMOTE_USER set")
+
+        if self.add_x_forwarded:
+            request_ip = self.request.META.get('REMOTE_ADDR')
+            self.log.debug("Proxy request IP: %s", request_ip)
+            request_headers['X-Forwarded-For'] = request_ip
+
+            request_proto = "https" if self.request.is_secure() else "http"
+            self.log.debug("Proxy request using %s", request_proto)
+            request_headers['X-Forwarded-Proto'] = request_proto
 
         return request_headers
 
     def get_quoted_path(self, path):
         """Return quoted path to be used in proxied request"""
         return quote_plus(path.encode('utf8'), QUOTE_SAFE)
+
+    def get_encoded_query_params(self):
+        """Return encoded query params to be used in proxied request"""
+        get_data = encode_items(self.request.GET.lists())
+        return urlencode(get_data)
 
     def _created_proxy_response(self, request, path):
         request_payload = request.body
@@ -149,8 +178,7 @@ class ProxyView(View):
         self.log.debug("Request URL: %s", request_url)
 
         if request.GET:
-            get_data = encode_items(request.GET.lists())
-            request_url += '?' + urlencode(get_data)
+            request_url += '?' + self.get_encoded_query_params()
             self.log.debug("Request URL: %s", request_url)
 
         try: