+++ CHANGELOG.rst
@@ -1,3 +1,12 @@
+0.10.0 (2020-02-05)
+===================
+
+* Fix add_remote_user when run without AuthenticationMiddleware. Fix #86
+* Add get_encoded_query_params method
+* Add support for Python 3.7 and 3.8.
+* Add support for Django 2.2 and 3.0.
+
+
 0.9.17 (2020-07-22)
 ===================